Fixed Flash Upload Problem

+++ irt_gui_5/src/main/java/irt/gui/controllers/flash/service/EraseObject.java
@@ -1,6 +1,6 @@
 package irt.gui.controllers.flash.service;
 public interface EraseObject {
 
-	boolean erase(long length);
+	boolean erase(int length);
 
 }
